ExtensionDiscovery: Clarify types and control flow when determining extension type from .info.yml file.
Somehow the IDE likes NULL more than FALSE as "empty" value for $file. Otherwise it complains further down, where it is used as an array index.

A condition with $file === NULL or $file === FALSE is generally preferable to empty($file) or !isset($file), because it makes the value range of a variable more transparent.